tests/test_clv/test_segmentation.py
import pytest
import numpy as np
import pandas as pd
from src.pipeline.clv import CustomerSegmentation
from src.pipeline.clv.base import BaseProcessor

class TestCustomerSegmentation:
    """Test suite for CustomerSegmentation"""

    def test_segmentation_initialization(self, config_loader):
        """Test segmentation initialization"""
        segmenter = CustomerSegmentation(config_loader)
        assert isinstance(segmenter, BaseProcessor)
        assert segmenter.segment_config is not None

    def test_process_data_interface(self, config_loader, sample_customer_features):
        """Test process_data method implementation"""
        segmenter = CustomerSegmentation(config_loader)
        result = segmenter.process_data(sample_customer_features)
        
        # Should return tuple of (DataFrame, Dict)
        assert isinstance(result, tuple)
        
        # Add more specific assertions
        processed_data, metadata = result
        assert isinstance(processed_data, pd.DataFrame)
        assert isinstance(metadata, dict)
        
        # Verify the processed data has required columns
        required_columns = [
            'customer_id', 
            'frequency', 
            'recency', 
            'monetary',
            'engagement_score',
            'engagement_level'
        ]
        assert all(col in processed_data.columns for col in required_columns)

    def test_rfm_segmentation(self, config_loader, sample_customer_features):
        """Test RFM segmentation"""
        segmenter = CustomerSegmentation(config_loader)
        df, model_data = segmenter.create_segments(sample_customer_features)
        
        # Check RFM scores exist
        assert 'R_score' in df.columns
        assert 'F_score' in df.columns
        assert 'M_score' in df.columns
        assert 'RFM_score' in df.columns
        
        # Check model data structure
        assert 'segment_ids' in model_data
        assert 'customer_features' in model_data

    def test_engagement_segmentation(self, config_loader, sample_customer_features):
        """Test engagement segmentation"""
        # Add engagement metrics
        data = sample_customer_features.copy()
        data['sms_active'] = np.random.randint(0, 2, len(data))
        data['email_active'] = np.random.randint(0, 2, len(data))
        
        segmenter = CustomerSegmentation(config_loader)
        df, _ = segmenter.create_segments(data)
        
        assert 'engagement_score' in df.columns
        assert 'engagement_level' in df.columns

    def test_segment_decoding(self, config_loader, sample_customer_features):
        """Test segment decoding functionality"""
        segmenter = CustomerSegmentation(config_loader)
        _, model_data = segmenter.create_segments(sample_customer_features)
        
        decoded = segmenter.decode_segments(model_data['segment_ids'])
        assert not decoded.empty

    def test_processor_interface_compliance(self, config_loader):
        """Test compliance with BaseProcessor interface"""
        segmenter = CustomerSegmentation(config_loader)
        
        # Check required methods
        assert hasattr(segmenter, 'process_data')
        
        # Check method signature
        from inspect import signature
        sig = signature(segmenter.process_data)
        assert 'df' in sig.parameters
        assert sig.return_annotation == pd.DataFrame

    def test_process_data_interface(self):
        # ... existing code ...
        
        # Replace the failing assert False with proper test logic
        test_data = {
            'customer_id': [1, 2, 3],
            'purchase_amount': [100, 200, 300],
            'frequency': [5, 3, 7]
        }
        segmentation = CustomerSegmentation()  # Assuming this is your class
        result = segmentation.process_data(test_data)
        
        # Add proper assertions
        assert result is not None
        assert isinstance(result, dict) or isinstance(result, pd.DataFrame)
        assert len(result) > 0
        # ... existing code ...